Enhance Zen Language Server with new features and refactor code structure
- Added constant declaration syntax to the language specification.
- Refactored LSP modules: moved types and utility functions to dedicated files for better organization.
- Introduced new LSP capabilities including hover, semantic tokens, call hierarchy, rename, signature help, inlay hints, code lens, and code actions.
- Improved document handling and background analysis in the Zen Language Server.
- Enhanced type inference functions and added support for inferring receiver types from the document store.
- Cleaned up unused imports and optimized existing code for better readability and maintainability.
